requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
torch==2.3.1
torchvision==0.18.1
opencv-python-headless==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.4.0
python-dotenv==1.0.1
